requirements.txt
# Core LangChain packages
langchain
langchain-community
langchain-google-genai
langgraph
langchain-pinecone
langchain-tavily
langchain-huggingface 
sentence-transformers

# Document loaders
pymupdf
python-docx
unstructured[docx,pptx]
pydantic
pydantic[email]

# Database
langgraph-checkpoint-postgres
sqlalchemy
psycopg2-binary
alembic

#Authentication
python_jose
PyJWT

# Utility packages
python-dotenv
loguru

# Apis
fastapi
uvicorn
